RM-1773_Ec2InstanceMachine_check_for_scheduled_events
* Add check for existing schedules to Ec2InstanceMachine before
  attempting to destroy them
* Update unavailable 'dawn' gem to dawnscanner

+++ lib/wonga/pantry/ec2_instance_machine.rb
@@ -101,7 +101,7 @@ module Wonga
         end
 
         after_transition to: :terminating do |machine, _state|
-          if machine.ec2_instance.instance_schedule
+          if machine.ec2_instance.instance_schedule && machine.ec2_instance.scheduled_event
             machine.ec2_instance.scheduled_event.destroy
           end
         end
